backend/requirements.txt
SQLAlchemy==2.0.44
numpy==2.3.4
pandas==2.3.3
psycopg2==2.9.11
pydantic==2.12.3
fastapi==0.121.0
python-dotenv==1.2.1